# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Librería Universal - API RESTful",
        "description": "API Backend profesional para el sistema de gestión de biblioteca y catálogo de Librería Universal. Desarrollado con Laravel, arquitectura en capas, JWT Authentication y control de acceso basado en roles.",
        "contact": {
            "name": "Equipo de Arquitectura Librería Universal",
            "email": "soporte@libreriauniversal.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor API Local / AI Studio"
        }
    ],
    "paths": {
...
        "/api/v1/books": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Listar catálogo de libros con paginación, filtros y búsqueda",
                "operationId": "61b0fdd35b8c14afc6e29eb90598649a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por título, autor o ISBN",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filtrar por nombre de categoría",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "available_only",
                        "in": "query",
                        "description": "Solo libros con copias disponibles (1/true)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordenación: latest, price-asc, price-desc, rating, title-asc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad por página (default 15, max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Colección paginada de libros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Libros obtenidos correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Book"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Libros"
                ],
                "summary": "Crear un nuevo libro en el catálogo (Requiere rol Admin o Bibliotecario)",
                "operationId": "5df8eb3d652dcc56b60cd15be3cbbf85",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "El Quijote de la Mancha"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Miguel de Cervantes"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Novelas"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25
                                    },
                                    "total_copies": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-8424116033"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Novela cumbre de la literatura en español."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Libro creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado (requiere rol admin o librarian)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/books/{id}": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Obtener detalles de un libro por su ID",
                "operationId": "c5b00bf4246c67b18e3092bd3665ea29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del libro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Libros"
                ],
                "summary": "Actualizar un libro existente (Requiere rol Admin o Bibliotecario)",
                "operationId": "3784164917f8ef14fad58d37b2f7de74",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Libro actualizado exitosamente"
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Libros"
                ],
                "summary": "Eliminar un libro del catálogo (Requiere rol Administrador)",
                "operationId": "0d28c2ad15a8fdbe85c2c95806b007f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libro eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Libro eliminado del catálogo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto: libro con préstamos activos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Libros"
                ],
                "summary": "Actualizar parcialmente un libro (Requiere rol Admin o Bibliotecario)",
                "operationId": "bd80aed78f62216a7bbbee191d4df41d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libro actualizado parcialmente"
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Book": {
                "title": "Libro",
                "description": "Entidad de libro en catálogo e inventario",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "numeric_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Cien Años de Soledad"
                    },
                    "author": {
                        "type": "string",
                        "example": "Gabriel García Márquez"
                    },
                    "category": {
                        "type": "string",
                        "example": "Novelas"
                    },
                    "isbn": {
                        "type": "string",
                        "example": "978-0307474728"
                    },
                    "description": {
                        "type": "string",
                        "example": "Obra cumbre del realismo mágico"
                    },
                    "publisher": {
                        "type": "string",
                        "example": "Editorial Sudamericana"
                    },
                    "year": {
                        "type": "integer",
                        "example": 1967
                    },
                    "pages": {
                        "type": "integer",
                        "example": 471
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 24.5
                    },
                    "originalPrice": {
                        "type": "number",
                        "format": "float",
                        "example": 28
                    },
                    "imageUrl": {
                        "type": "string",
                        "example": "https://images.unsplash.com/..."
                    },
                    "coverTheme": {
                        "type": "string",
                        "example": "emerald"
                    },
                    "stock": {
                        "type": "integer",
                        "example": 18
                    },
                    "available_copies": {
                        "type": "integer",
                        "example": 18
                    },
                    "total_copies": {
                        "type": "integer",
                        "example": 20
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.9
                    },
                    "reviewsCount": {
                        "type": "integer",
                        "example": 384
                    },
                    "featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "bestSeller": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación mediante JWT Bearer token. Ejemplo: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}